refactor: import custom hooks from hooks index
