+++ fableous-fe/src/containers/HubCanvasPage.tsx
@@ -21,11 +21,10 @@ import AchievementButton from "../components/achievement/AchievementButton";
 import Canvas from "../components/canvas/Canvas";
 import CursorScreen, { Cursor } from "../components/canvas/CursorScreen";
 import FormikTextField from "../components/FormikTextField";
-import { useAchievement, useWsConn } from "../hooks";
+import { useAchievement, useWsConn, useContainRatio } from "../hooks";
 import { ROLE_ICON, StudentRole } from "../constant";
 import BackButton from "../components/BackButton";
 import { ImperativeCanvasRef, TextShapeMap } from "../components/canvas/data";
-import useContainRatio from "../hooks/useContainRatio";
 import { ASPECT_RATIO } from "../components/canvas/constants";
 import ChipRow from "../components/ChipRow";
 import FormikTagField from "../components/FormikTagField";
