+++ fableous-fe/src/hooks/index.ts
@@ -1,4 +1,6 @@
 import useAchievement from "./useAchievement";
+import useContainRatio from "./useContainRatio";
+import useTutorial from "./useTutorial";
 import useWsConn from "./useWsConn";
 
-export { useAchievement, useWsConn };
+export { useAchievement, useContainRatio, useTutorial, useWsConn };
